+++ packages/shared-components/src/questions/SCAnswerOptions.tsx
@@ -9,7 +9,7 @@ import ChoiceFeedback from '../evaluation/ChoiceFeedback'
 
 export interface SCAnswerOptionsProps {
   displayMode?: ElementDisplayMode
-  choices: Partial<Choice>[]
+  choices: Choice[]
   feedbacks?: QuestionFeedback[] | null
   value?: Record<number, boolean>
   onChange: (value: Record<number, boolean>) => void
@@ -45,9 +45,9 @@ export function SCAnswerOptions({
           b: (text) => <span className="font-bold">{text}</span>,
         })}
       </div>
-      {choices.map((choice, index) => {
+      {choices.map((choice) => {
         return (
-          <div key={`sc-choice-${index}-${choice.value}`}>
+          <div key={`sc-choice-${choice.ix}-${choice.value}`}>
             <Button
               fluid
               className={{
@@ -55,17 +55,21 @@ export function SCAnswerOptions({
                   'hover:bg-unset min-h-[2.5rem] border-slate-400',
                   !hideFeedbacks &&
                     feedbacks &&
-                    feedbacks[index] &&
+                    feedbacks[choice.ix] &&
                     'rounded-b-none'
                 ),
               }}
               onClick={() =>
                 onChange(
-                  Object.fromEntries(choices.map((_, i) => [i, i === index]))
+                  Object.fromEntries(
+                    choices.map((_, i) => [i, i === choice.ix])
+                  )
                 )
               }
-              active={value?.[index]}
-              data={{ cy: `sc-${elementIx + 1}-answer-option-${index + 1}` }}
+              active={value?.[choice.ix]}
+              data={{
+                cy: `sc-${elementIx + 1}-answer-option-${choice.ix + 1}`,
+              }}
               disabled={disabled}
             >
               <Button.Label>
@@ -78,8 +82,8 @@ export function SCAnswerOptions({
                 />
               </Button.Label>
             </Button>
-            {!hideFeedbacks && feedbacks && feedbacks[index] && (
-              <ChoiceFeedback feedback={feedbacks[index]!} />
+            {!hideFeedbacks && feedbacks && feedbacks[choice.ix] && (
+              <ChoiceFeedback feedback={feedbacks[choice.ix]!} />
             )}
           </div>
         )